print timestamps using the preferred formatting of the current locale
Requires using the dateutil module in order to set a sane timezone on
the date object, requires setting LC_TIME (done in a preceding commit)
in order to have functionally locale-aware strftime formatters.

Alternative fix which uses native python instead of subprocesses, at the
cost of an extra external dependency.